+++ packages/apple_vision_image_classification/ios/Classes/AppleVisionImageClassificationPlugin.swift
@@ -1,0 +1,1 @@
+/Users/johnsparkman/Documents/GitHub/apple_vision/packages/apple_vision_face/macos/Classes/AppleVisionFacePlugin.swift