+++ packages/apple_vision_object/ios/Classes/AppleVisionObject.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface AppleVisionObject : NSObject<FlutterPlugin>
+@end