+++ packages/apple_vision_recognize_text/ios/Classes/AppleVisionRecognizeText.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface AppleVisionRecognizeText : NSObject<FlutterPlugin>
+@end