+++ packages/apple_vision_recognize_text/macos/Classes/AppleVisionRecognizeTextPlugin.swift
@@ -1,15 +1,26 @@
 import AVFoundation
-import FlutterMacOS
 import Vision
+
+#if os(iOS)
+import Flutter
+import UIKit
+#elseif os(macOS)
+import FlutterMacOS
 import AppKit
+#endif
 
 public class AppleVisionRecognizeTextPlugin: NSObject, FlutterPlugin {
     
     let registry: FlutterTextureRegistry
     
     public static func register(with registrar: FlutterPluginRegistrar) {
-        let instance = AppleVisionRecognizeTextPlugin(registrar.textures)
+        #if os(iOS)
+        let method = FlutterMethodChannel(name:"apple_vision/recognize_text", binaryMessenger: registrar.messenger())
+        let instance = AppleVisionFacePlugin(registrar.textures())
+        #elseif os(macOS)
         let method = FlutterMethodChannel(name:"apple_vision/recognize_text", binaryMessenger: registrar.messenger)
+        let instance = AppleVisionFacePlugin(registrar.textures)
+        #endif
         registrar.addMethodCallDelegate(instance, channel: method)
     }