+++ packages/apple_vision/CHANGELOG.md
@@ -1,6 +1,13 @@
 ## 0.0.2
 
 * Added ios compatibility, and updated comments.
+* Added Animal Pose
+* Added Object Classification
+* Added Image Classification
+* Added Face Detection
+* Added Recognize Text
+* Added Selfie Segmentation
+* Added Object Tracking
 
 ## 0.0.1